use size_t for both

+++ src/stan/services/sample/standalone_gqs.hpp
@@ -47,13 +47,13 @@ void get_model_parameters(const Model &model,
   size_t num_params = param_names.size();
   std::vector<std::vector<size_t>> dimss;
   model.get_dims(dimss);
-  for (int i = 0;i < param_names.size(); i++) {
+  for (size_t i = 0;i < param_names.size(); i++) {
     for (size_t j = i; j < all_param_names.size(); ++j) {
       if (param_names[i].compare(all_param_names[j]) == 0) {
         param_dimss.emplace_back(dimss[j]);
         break;
       }
-    }   
+    }
   }
 }
